+++ NEWS.md
@@ -1,3 +1,12 @@
+# isogeochem 1.0.2
+
+* Updated the DESCRIPTION file
+* Updated the README file
+* Updated the isogeochem_manual.pdf file
+* Update to temp_D48() #1: the curve_intersect() part was slimmed down
+* Update to temp_D48() #2: temperatures are now calculated with the "Fiebig21" eq
+* Typo corrected in a18_H2O_OH(): "Z21-X3LYP" —> "Z20-X3LYP"
+
 # isogeochem 1.0.1
 
 * This is the first stable release of isogeochem.